colocate auth components next to router
