+++ frontend/router/components/AuthAnyAdminRoutes/AuthAnyAdminRoutes.tsx
@@ -2,24 +2,22 @@ import React, { useContext } from "react";
 import { useErrorHandler } from "react-error-boundary";
 import { AppContext } from "context/app";
 
-interface IAuthenticatedAdminRoutesProps {
+interface IAuthAnyAdminRoutesProps {
   children: JSX.Element;
 }
 
 /**
- * Checks if a user is any maintainer or any admin when routing
+ * Checks if a user is any global or team admin when routing
  */
-const AuthenticatedAdminRoutes = ({
-  children,
-}: IAuthenticatedAdminRoutesProps): JSX.Element | null => {
+const AuthAnyAdminRoutes = ({ children }: IAuthAnyAdminRoutesProps) => {
   const handlePageError = useErrorHandler();
-  const { currentUser, isGlobalAdmin } = useContext(AppContext);
+  const { currentUser, isGlobalAdmin, isAnyTeamAdmin } = useContext(AppContext);
 
   if (!currentUser) {
     return null;
   }
 
-  if (!isGlobalAdmin) {
+  if (!isGlobalAdmin && !isAnyTeamAdmin) {
     handlePageError({ status: 403 });
     return null;
   }
@@ -27,4 +25,4 @@ const AuthenticatedAdminRoutes = ({
   return <>{children}</>;
 };
 
-export default AuthenticatedAdminRoutes;
+export default AuthAnyAdminRoutes;
