+++ frontend/router/components/AuthenticatedRoutes/AuthenticatedRoutes.tsx
