+++ frontend/router/components/AuthenticatedRoutes/index.ts
