+++ frontend/router/index.tsx
@@ -3,7 +3,6 @@ import {
   browserHistory,
   IndexRedirect,
   IndexRoute,
-  InjectedRouter,
   Route,
   RouteComponent,
   Router,
@@ -15,12 +14,6 @@ import AdminUserManagementPage from "pages/admin/UserManagementPage";
 import AdminTeamManagementPage from "pages/admin/TeamManagementPage";
 import TeamDetailsWrapper from "pages/admin/TeamManagementPage/TeamDetailsWrapper";
 import App from "components/App";
-import AuthenticatedAdminRoutes from "components/AuthenticatedAdminRoutes";
-import AuthAnyAdminRoutes from "components/AuthAnyAdminRoutes";
-import AuthenticatedRoutes from "components/AuthenticatedRoutes";
-import UnauthenticatedRoutes from "components/UnauthenticatedRoutes";
-import AuthGlobalAdminMaintainerRoutes from "components/AuthGlobalAdminMaintainerRoutes";
-import AuthAnyMaintainerAnyAdminRoutes from "components/AuthAnyMaintainerAnyAdminRoutes";
 import ConfirmInvitePage from "pages/ConfirmInvitePage";
 import ConfirmSSOInvitePage from "pages/ConfirmSSOInvitePage";
 import CoreLayout from "layouts/CoreLayout";
@@ -55,6 +48,14 @@ import AgentOptionsPage from "pages/admin/TeamManagementPage/TeamDetailsWrapper/
 import PATHS from "router/paths";
 import AppProvider from "context/app";
 import RoutingProvider from "context/routing";
+
+import AuthGlobalAdminRoutes from "./components/AuthGlobalAdminRoutes";
+import AuthAnyAdminRoutes from "./components/AuthAnyAdminRoutes";
+import AuthenticatedRoutes from "./components/AuthenticatedRoutes";
+import UnauthenticatedRoutes from "./components/UnauthenticatedRoutes";
+import AuthGlobalAdminMaintainerRoutes from "./components/AuthGlobalAdminMaintainerRoutes";
+import AuthAnyMaintainerAnyAdminRoutes from "./components/AuthAnyMaintainerAnyAdminRoutes";
+import PremiumRoutes from "./components/PremiumRoutes";
 
 interface IAppWrapperProps {
   children: JSX.Element;
@@ -101,7 +102,7 @@ const routes = (
           <Route path="settings" component={AuthAnyAdminRoutes}>
             <IndexRedirect to={"/dashboard"} />
             <Route component={SettingsWrapper}>
-              <Route component={AuthenticatedAdminRoutes}>
+              <Route component={AuthGlobalAdminRoutes}>
                 <Route path="organization" component={AdminAppSettingsPage} />
                 <Route
                   path="organization/:section"
@@ -109,7 +110,9 @@ const routes = (
                 />
                 <Route path="integrations" component={AdminIntegrationsPage} />
                 <Route path="users" component={AdminUserManagementPage} />
-                <Route path="teams" component={AdminTeamManagementPage} />
+                <Route component={PremiumRoutes}>
+                  <Route path="teams" component={AdminTeamManagementPage} />
+                </Route>
               </Route>
             </Route>
             <Route path="teams/:team_id" component={TeamDetailsWrapper}>
